backend/requirements.txt
fastapi 
uvicorn 
boto3 
AWS 
python-dotenv 
pydantic